Create Card and Customer repository and change creditLimit data type to BigDecimal

+++ src/test/java/com/example/card_management_system/Entity/CardTest.java
@@ -3,6 +3,8 @@ package com.example.card_management_system.Entity;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class CardTest {
@@ -12,8 +14,8 @@ class CardTest {
     void setCreditLimit() {
 
         Card card = new Card();
-        card.setCreditLimit(20.0);
-        assertEquals(20.0,card.getCreditLimit());
+        card.setCreditLimit(BigDecimal.valueOf(20.0));
+        assertEquals(BigDecimal.valueOf(20.0),card.getCreditLimit());
 
     }
